feat: add StairNet

+++ util/util.py
@@ -69,7 +69,7 @@ def tensor2im(input_image, imtype=np.uint8):
         if image_numpy.shape[0] == 1:  # grayscale to RGB
             image_numpy = np.tile(image_numpy, (3, 1, 1))
         image_numpy = np.transpose(image_numpy, (1, 2, 0))  * 255  # post-processing: tranpose and scaling
-        input_image = torch.clamp(input_image, 0, 255)
+        input_image = torch.clamp(input_image.type(torch.float32), 0, 255)
     else:  # if it is a numpy array, do nothing
         image_numpy = input_image
     return image_numpy.astype(imtype)
